health_fitness_app_package/backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
pymongo==4.3.3
python-dotenv==1.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.4.4
gunicorn==20.1.0
scikit-learn==1.2.2
numpy==1.24.2
pandas==1.5.3
